requirements.txt
# XAUUSD Gold Trading System - Python Dependencies
# Python 3.11+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async Support
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1

# WebSocket
websockets==12.0
python-socketio==5.11.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# MT5 Integration (Windows only)
MetaTrader5==5.0.45; sys_platform == 'win32'

# Telegram
# python-telegram-bot==20.7
# python-telegram-bot[job-queue]==20.7

# Data Processing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Technical Analysis
# ta-lib==0.4.28

# Logging
structlog==24.1.0
python-json-logger==2.0.7
colorlog==6.8.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# HTTP Client
# httpx==0.26.0
requests==2.31.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Task Queue
celery==5.3.4
kombu==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
arrow==1.3.0

# Validation
email-validator==2.1.0
phonenumbers==8.13.27

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.0.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Performance
ujson==5.9.0
orjson==3.9.10

# Retry Logic
tenacity==8.2.3
backoff==2.2.1

# Rate Limiting
slowapi==0.1.9

# Environment
python-decouple==3.8

# Type Hints
types-redis==4.6.0.20240106
types-requests==2.31.0.20240106
types-PyYAML==6.0.12.12

# Scheduling
apscheduler==3.10.4

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# File Handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Compression
zstandard==0.22.0

# System Monitoring
psutil==5.9.7

# Timezone
tzdata==2023.4
PyJWT==2.8.0